% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{.onLoad}
\alias{.onLoad}
\title{Package Startup}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{The library path where the package is installed.}

\item{pkgname}{The name of the package.}
}
\description{
Automatically loads the default 15-topic models (DFM and KMeans) along with cluster labels
when the package is loaded.
}
\keyword{internal}
