% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_topic.R
\name{find_topic}
\alias{find_topic}
\title{Find Topics Across Models}
\usage{
find_topic(topic_keyword, pkgname = utils::packageName(), verbose = FALSE)
}
\arguments{
\item{topic_keyword}{A keyword to search for in cluster labels.}

\item{pkgname}{The name of the package. Default is the current package name.}

\item{verbose}{Logical. If TRUE, prints debug information. Default is FALSE.}
}
\value{
A list of models containing the topic of interest.
}
\description{
Find Topics Across Models
}
\examples{
\dontrun{
find_topic("Cybersecurity")
}

}
