% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_topics.R
\name{list_all_topics}
\alias{list_all_topics}
\title{Extract Topics from All Models}
\usage{
list_all_topics(pkgname = utils::packageName())
}
\arguments{
\item{pkgname}{The package name (default is the current package).}
}
\value{
A data frame containing topics and the associated models.
}
\description{
Extracts all unique topics from the available KMeans models in the package.
}
\examples{
SectorinsightRv2::list_all_topics()
}
