% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_models.R
\name{list_models}
\alias{list_models}
\title{List All Available Models}
\usage{
list_models(pkgname = utils::packageName())
}
\arguments{
\item{pkgname}{The package name (default is the current package).}
}
\value{
A data frame containing model names, number of topics, and file paths.
}
\description{
Returns a list of all available KMeans models in the package along with their metadata (e.g., number of topics).
}
\examples{
SectorinsightRv2::list_models()
}
