% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{load_model}
\alias{load_model}
\title{Load a Pre-Trained Model}
\usage{
load_model(topics, pkgname = utils::packageName())
}
\arguments{
\item{topics}{An integer specifying the number of topics (e.g., 15, 25).}

\item{pkgname}{The package name (default is the current package name).}
}
\value{
Loads the models and assigns them to the package's namespace:
\itemize{
\item \code{dfm_model}: The loaded DFM model.
\item \code{kmeans_model}: The loaded KMeans model.
\item \code{cluster_labels_df}: A data frame containing cluster IDs and their suggested labels.
}
}
\description{
Dynamically loads a pre-trained DFM and KMeans model for a specified number of topics.
It also extracts cluster labels from the KMeans model and annotates the cluster assignments.
}
\examples{
\dontrun{
# Load the 25-topic model
load_model(topics = 25)
}

}
